Optimizer: convert the SILPrinter and  TestInstructionIteration passes to tests
