+++ SwiftCompilerSources/Sources/Optimizer/PassManager/PassRegistration.swift
@@ -148,8 +148,6 @@ private func registerSwiftPasses() {
   registerForSILCombine(EndCOWMutationAddrInst.self, { run(EndCOWMutationAddrInst.self, $0) })
 
   // Test passes
-  registerPass(silPrinterPass, { silPrinterPass.run($0) })
-  registerPass(testInstructionIteration, { testInstructionIteration.run($0) })
   registerPass(updateBorrowedFromPass, { updateBorrowedFromPass.run($0) })
 }
