+++ SwiftCompilerSources/Sources/Optimizer/TestPasses/CMakeLists.txt
@@ -7,7 +7,5 @@
 # See http://swift.org/CONTRIBUTORS.txt for Swift project authors
 
 swift_compiler_sources(Optimizer
-  SILPrinter.swift
   UpdateBorrowedFrom.swift
-  TestInstructionIteration.swift
 )
