+++ SwiftCompilerSources/Sources/SIL/Utilities/Test.swift
@@ -142,6 +142,8 @@ public func registerTests() {
   // Register each test.
   registerTests(
     parseTestSpecificationTest,
+    basicBlockTest,
+    instructionIterationTest,
     smallProjectionPathTest,
     getAccessBaseTest,
     accessPathTest,
